Chore: Improve the data model and performance of the state sync
