+++ sqlmesh/core/snapshot/definition.py
@@ -1188,6 +1188,10 @@ class Snapshot(PydanticModel, SnapshotInfoMixin):
         return to_timestamp(self.ttl, relative_base=to_datetime(self.updated_ts))
 
     @property
+    def ttl_ms(self) -> int:
+        return self.expiration_ts - self.updated_ts
+
+    @property
     def custom_materialization(self) -> t.Optional[str]:
         if self.is_custom:
             return t.cast(CustomKind, self.model.kind).materialization
